+++ GUI Sources/frontend/wailsjs/go/main/App.d.ts
@@ -39,6 +39,14 @@ export function RldevOrgTextImport(arg1:string,arg2:string,arg3:string,arg4:stri
 
 export function RldevPngToG00(arg1:string,arg2:string,arg3:string,arg4:string,arg5:boolean):Promise<string>;
 
+export function RldevSaveDump(arg1:string,arg2:boolean,arg3:boolean):Promise<string>;
+
+export function RldevSaveGet(arg1:string,arg2:string):Promise<string>;
+
+export function RldevSaveInfo(arg1:string):Promise<string>;
+
+export function RldevSaveSet(arg1:string,arg2:string,arg3:boolean):Promise<string>;
+
 export function RldevXmlToGan(arg1:string,arg2:string):Promise<string>;
 
 export function SelectDirectory(arg1:string):Promise<string>;